Changes required for test environment.  DROP THIS FROM ANY PULL REQUESTS MADE TO usegalaxy-au/usegalaxy-au-tools.
Changes required for test environment.  DROP THIS FROM ANY PULL REQUESTS MADE TO usegalaxy-au/usegalaxy-au-tools.

.env and .gitgnore changes required for this fork

This fork needs cat-dev and galaxy-cat urls in .env for testing

Not ignoring cat-dev on this fork

in the test environment, lets make sure we are using the test URLs before we install anything

comment out hub pull-request, hub only allows one set of credentials and we do not want the robot to be making pull requests to cat-bro fork